% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakAnnotation.R
\name{AnnoUpsetPlot}
\alias{AnnoUpsetPlot}
\title{AnnoUpsetPlot}
\usage{
AnnoUpsetPlot(LinkageObject, order_by = "freq")
}
\arguments{
\item{LinkageObject}{An LinkageObject after the annottation.}

\item{order_by}{How the intersections in the matrix should be ordered by. Options include frequency (entered as "freq"), degree, or both in any order.}
}
\value{
An upsetplot after the annottation.
}
\description{
AnnoUpsetPlot
}
\examples{
library(Linkage)
data("SmallLinkageObject")
gene_list <- c("TSPAN6", "CD99", "KLHL13")
LinkageObject <-
  RegulatoryPeak(
    LinkageObject = SmallLinkageObject,
    gene_list = gene_list,
    genelist_idtype = "external_gene_name"
  )
peakAnno <- PeakAnnottation(LinkageObject, Species = "Homo")
AnnoUpsetPlot(LinkageObject = peakAnno)
}
