% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNetwork.R
\name{BuildNetwork}
\alias{BuildNetwork}
\title{BuildNetwork.}
\usage{
BuildNetwork(LinkageObject)
}
\arguments{
\item{LinkageObject}{An Linkage Object after CreateNetworkObject.}
}
\value{
An visNetwork
}
\description{
BuildNetwork.
}
\examples{
RNA.seq.dir <- system.file("extdata", "TCGA-BRCA-RNA.txt", package = "Linkage")
RNA.seq <- data.table::fread(RNA.seq.dir, header = TRUE)
Homo.list.files.dir <- system.file("extdata", "Homo.ATAC", package = "Linkage")
Homo.list.files <- list.files(Homo.list.files.dir)
Homo.list.files <- paste0(Homo.list.files.dir, "/", Homo.list.files)
Homo.df_list <- lapply(Homo.list.files, function(file) data.table::fread(file, header = TRUE))
ATAC.seq <- do.call(rbind, Homo.df_list)
LinkageObject <- CreateLinkageObject(ATAC_count = ATAC.seq, RNA_count = RNA.seq, Species = "Homo", id_type = "ensembl_gene_id")
gene_list <- c("TSPAN6", "CD99", "KLHL13", "ARX", "HCCS")
LinkageObject <- regulatory_peak(LinkageObject = LinkageObject, gene_list = gene_list, genelist_idtype = "external_gene_name")
LinkageObject <- MultipleMotifAnalysis(LinkageObject = LinkageObject,Species = "Homo",TF_cor_method = "pearson")
LinkageObject <- CreateNetworkObject(LinkageObject = LinkageObject, genelist_idtype = "entrezgene_id", filter_col = "FDR", filter_value = 0.01)
BuildNetwork(LinkageObject)
}
