% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegulatoryPeaks.R
\name{CorrPlot}
\alias{CorrPlot}
\title{Correlation scatter plot.}
\usage{
CorrPlot(LinkageObject, gene, color = "black", fill = "lightgray")
}
\arguments{
\item{LinkageObject}{An Linkage Object.}

\item{gene}{A gene that want to be visualized.}

\item{color}{The color of the straight line.}

\item{fill}{The color of the fill.}
}
\value{
Scatterplot of correlation between genes and peaks.
}
\description{
Correlation scatter plot.
}
\examples{
library(Linkage)
data("SmallLinkageObject")
gene_list <- c("TSPAN6", "CD99", "KLHL13")
LinkageObject <-
  RegulatoryPeak(
    LinkageObject = SmallLinkageObject,
    gene_list = gene_list,
    genelist_idtype = "external_gene_name"
  )
CorrPlot(LinkageObject, gene = "TSPAN6")
}
