% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateLinkageObject.R
\name{CreateLinkageObject}
\alias{CreateLinkageObject}
\title{Create a Linkage Object}
\usage{
CreateLinkageObject(ATAC_count, RNA_count, Species, id_type)
}
\arguments{
\item{ATAC_count}{The chromatin accessibility matrix file is a tab-delimited multi-column data matrix as well, which the first three columns represent chromosome name, start coordinate on the chromosome and end coordinate on the chromosome of the peaks respectively; the remaining columns of the chromatin accessibility matrix file represent normalized or raw chromatin accessibility levels of peaks for each sample.}

\item{RNA_count}{The gene expression matrix file is a tab-delimited multi-column data matrix, which the first column represents gene symbols and the following columns represent normalized or raw expression levels of genes for each sample.}

\item{Species}{Select the species, Homo or Mus.}

\item{id_type}{The RNA_count gene id type. There are ensembl_gene_id, external_gene_name, entrezgene_id types of genetic IDs to choose from.}
}
\value{
A Linkage Object
}
\description{
Create a Linkage Object
}
\examples{
library(LinkageR)
LinkageObject <-
  CreateLinkageObject(
    ATAC_count = Small_ATAC,
    RNA_count = Small_RNA,
    Species = "Homo",
    id_type = "ensembl_gene_id"
  )
}
