% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.KEGG.Enrichment.R
\name{EnrichBarPlot}
\alias{EnrichBarPlot}
\title{Enrichment analysis Barplot}
\usage{
EnrichBarPlot(object, size = 8)
}
\arguments{
\item{object}{Go or KEGG object.}

\item{size}{font size.}
}
\value{
Enrichment bar plot.
}
\description{
Enrichment analysis Barplot
}
\examples{
library(linkage)
library(LinkageData)
Geneset <- system.file("extdata", "SASPGeneSet.rdata", package = "LinkageData")
load(Geneset)
go <-
  GOEnrichment(gene_list = SASPGeneSet,
               Species = "Homo",
               genelist_idtype = "external_gene_name")
p <- EnrichBarPlot(go)
}
