% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.KEGG.Enrichment.R
\name{EnrichCnetPlot}
\alias{EnrichCnetPlot}
\title{Enrichment analysis Cnetplot}
\usage{
EnrichCnetPlot(object, size = 8)
}
\arguments{
\item{object}{Go or KEGG object}

\item{size}{font size.}
}
\value{
Enrichment network.
}
\description{
Enrichment analysis Cnetplot
}
\examples{
library(Linkage)
data("SASPGeneSet")
go <-
  GOEnrichment(gene_list = SASPGeneSet,
               Species = "Homo",
               genelist_idtype = "external_gene_name")
p <- EnrichCnetPlot(go)
}
