% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.KEGG.Enrichment.R
\name{EnrichUpsetPlot}
\alias{EnrichUpsetPlot}
\title{Enrichment analysis Upsetplot}
\usage{
EnrichUpsetPlot(object, size = 8)
}
\arguments{
\item{object}{Go or KEGG object}

\item{size}{font size.}
}
\value{
Enrichment upset plot.
}
\description{
Enrichment analysis Upsetplot
}
\examples{
library(LinkageR)
data("SASPGeneSet")
go <-
  GOEnrichment(gene_list = SASPGeneSet,
               Species = "Homo",
               genelist_idtype = "external_gene_name")
p <- EnrichUpsetPlot(go)
}
