% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNetwork.R
\name{FilterGRNs}
\alias{FilterGRNs}
\title{Create Network Object.}
\usage{
FilterGRNs(LinkageObject, genelist_idtype, filter_col, filter_value)
}
\arguments{
\item{LinkageObject}{An Linkage Object after BuildGRNs.}

\item{genelist_idtype}{There are ensembl_gene_id, external_gene_name, entrezgene_id types of genetic IDs to choose from.}

\item{filter_col}{p_value,FDR,rho can be selected.}

\item{filter_value}{Cutoff value of filter_col.}
}
\value{
Network Object.
}
\description{
Create Network Object.
}
\examples{
library(LinkageR)
data("Small_ATAC.rda")
data("Small_RNA.rda")
LinkageObject <-
  CreateLinkageObject(
    ATAC_count = Small_ATAC,
    RNA_count = Small_RNA,
    Species = "Homo",
    id_type = "ensembl_gene_id"
  )
gene_list <- c("TSPAN6", "CD99", "KLHL13")
LinkageObject <-
  RegulatoryPeak(
    LinkageObject = LinkageObject,
    gene_list = gene_list,
    genelist_idtype = "external_gene_name"
  )
LinkageObject <-
  BuildGRNs(LinkageObject = LinkageObject,
                        Species = "Homo",
                        TF_cor_method = "pearson")
LinkageObject <-
  FilterGRNs(
    LinkageObject = LinkageObject,
    genelist_idtype = "entrezgene_id",
    filter_col = "FDR",
    filter_value = 0.01
  )
}
