% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Homo.position}
\alias{Homo.position}
\title{Homo sapiens genetic coordinates}
\format{
## `Homo.position`
A data frame with 66,206 rows and 6 columns:
\describe{
  \item{ensembl_gene_id, external_gene_name, entrezgene_id}{Gene ID type}
  \item{chromosome_name}{chromosome name}
  \item{start_position}{chromosome start site}
  \item{end_position}{chromosome termination site}
}

## `Homo.position`
A data frame with 66,206 rows and 6 columns:
\describe{
  \item{ensembl_gene_id, external_gene_name, entrezgene_id}{Gene ID type}
  \item{chromosome_name}{chromosome name}
  \item{start_position}{chromosome start site}
  \item{end_position}{chromosome termination site}
}
}
\usage{
Homo.position

Homo.position
}
\description{
Used to find the coordinates of each gene
Report ...

Used to find the coordinates of each gene
Report ...
}
\keyword{datasets}
