% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Install.packge.R
\name{Install.biocViews}
\alias{Install.biocViews}
\title{Used when the linkage is installed for the first time and is used to install dependencies}
\usage{
Install.biocViews()
}
\value{
Install the dependencies under bioconductor
}
\description{
Used when the linkage is installed for the first time and is used to install dependencies
}
\examples{
Linkage::Install.biocViews()
}
