% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunExample.R
\name{Linkage}
\alias{Linkage}
\title{Launch Shiny App}
\usage{
Linkage()
}
\arguments{
\item{name}{The name of the app to run}

\item{...}{arguments to pass to shiny::runApp}
}
\description{
Launch Shiny App
}
