% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateLinkageObject.R
\docType{class}
\name{LinkageObject-class}
\alias{LinkageObject-class}
\title{Linkage Class}
\value{
Linkage Class
}
\description{
Linkage Class
}
\section{Slots}{

\describe{
\item{\code{RNA.mtrix}}{RNA-seq count.}

\item{\code{ATAC.matrix}}{ATAC-seq count.}

\item{\code{active.gene}}{selected genes.}

\item{\code{cor.peak}}{Regulatory peaks for each selected genes.}

\item{\code{cor.peak.annotation}}{Regulatory peaks annotation.}

\item{\code{detailpeakannotation}}{Regulatory peaks annotation detail.}

\item{\code{Motif}}{The all TFs of selected genes.}

\item{\code{Gene_TF}}{The hight correlation TFs of selected genes.}

\item{\code{geneid}}{The selcted genes id.}

\item{\code{Summary}}{Summary}
}}

