% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAnalysis.R
\name{motif_analysis}
\alias{motif_analysis}
\title{Regulatory peak motif enrichment analysis}
\usage{
motif_analysis(peakfile, Species)
}
\arguments{
\item{peakfile}{ATAC-Seq expression matrix or bed file}

\item{Species}{Select the species, Homo or Mus.}
}
\value{
Motif result
}
\description{
Regulatory peak motif enrichment analysis
}
\examples{
dir <- system.file("extdata","ENSG00000000419.csv", package = "Linkage")
peakfile <- read.csv(dir, header = T)
motif_analysis(peakfile, "Homo")
}
