% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAnalysis.R
\name{MotifEnrichment}
\alias{MotifEnrichment}
\title{Regulatory peak motif enrichment analysis}
\usage{
MotifEnrichment(peakfile, Species)
}
\arguments{
\item{peakfile}{ATAC-Seq expression matrix or bed file}

\item{Species}{Select the species, Homo or Mus.}
}
\value{
Motif result
}
\description{
Regulatory peak motif enrichment analysis
}
\examples{
library(linkage)
library(LinkageData)
peakpath <- system.file("extdata","ENSG00000000419.rdata",package = "LinkageData")
load(peakpath)
motif <- MotifEnrichment(PeakFile, "Homo")
head(motif)
}
