% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAnalysis.R
\name{MultipleMotifAnalysis}
\alias{MultipleMotifAnalysis}
\title{MotifAnalysis.}
\usage{
MultipleMotifAnalysis(LinkageObject, Species = "Homo", TF_cor_method)
}
\arguments{
\item{LinkageObject}{An Linkage Object after regulatory_peak.}

\item{Species}{Select the species, Homo or Mus.The default is Homo.}

\item{TF_cor_method}{For correlation calculation, pearson/spearman/kendall can be selected.}
}
\value{
An Linkage Object after MultipleMotifAnalysis
}
\description{
MotifAnalysis.
}
\examples{
RNA.seq.dir <- system.file("extdata", "TCGA-BRCA-RNA.txt", package = "Linkage")
RNA.seq <- data.table::fread(RNA.seq.dir, header = TRUE)
Homo.list.files.dir <- system.file("extdata", "Homo.ATAC", package = "Linkage")
Homo.list.files <- list.files(Homo.list.files.dir)
Homo.list.files <- paste0(Homo.list.files.dir, "/", Homo.list.files)
Homo.df_list <- lapply(Homo.list.files, function(file) data.table::fread(file, header = TRUE))
ATAC.seq <- do.call(rbind, Homo.df_list)
LinkageObject <- CreateLinkageObject(ATAC_count = ATAC.seq, RNA_count = RNA.seq, Species = "Homo", id_type = "ensembl_gene_id")
gene_list <- c("TSPAN6", "CD99", "KLHL13", "ARX", "HCCS")
LinkageObject <- regulatory_peak(LinkageObject = LinkageObject, gene_list = gene_list, genelist_idtype = "external_gene_name")
LinkageObject <- MultipleMotifAnalysis(LinkageObject = LinkageObject,Species = "Homo",TF_cor_method = "pearson")
}
