% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mus.PFMatrixList}
\alias{Mus.PFMatrixList}
\title{Mus musculus sapiens PWM matrix list of motif}
\format{
## `Mus.PFMatrixList`
}
\usage{
Mus.PFMatrixList
}
\description{
include 144 motifs
}
\details{
Report ...
}
\keyword{datasets}
