% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mus.position}
\alias{Mus.position}
\title{Mus musculus sapiens genetic coordinates}
\format{
## `Mus.position`
A data frame with 57,451 rows and 6 columns:
\describe{
  \item{ensembl_gene_id, external_gene_name, entrezgene_id}{Gene ID type}
  \item{chromosome_name}{chromosome name}
  \item{start_position}{chromosome start site}
  \item{end_position}{chromosome termination site}
}
}
\usage{
Mus.position
}
\description{
Used to find the coordinates of each gene
Report ...
}
\keyword{datasets}
