% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mus.pwm_library_dt}
\alias{Mus.pwm_library_dt}
\title{Mus musculus sapiens motif id and name}
\format{
## `Mus.pwm_library_dt`
A data frame with 144 rows and 2 columns:
\describe{
  \item{ID}{motif id}
  \item{name}{motif name}
}
}
\usage{
Mus.pwm_library_dt
}
\description{
Report ...
}
\keyword{datasets}
