% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakAnnotation.R
\name{PeakAnnottation}
\alias{PeakAnnottation}
\title{Annotation regulatory peak}
\usage{
PeakAnnottation(LinkageObject, Species = "Homo")
}
\arguments{
\item{LinkageObject}{An Linkage Object after RegulatoryPeak.}

\item{Species}{Select the species, Homo or Mus.The default is Homo.}
}
\value{
An LinkageObject after the annottation.
}
\description{
Annotation regulatory peak
}
\examples{
library(linkage)
data("SmallLinkageObject")
gene_list <- c("TSPAN6", "CD99", "KLHL13")
LinkageObject <-
  RegulatoryPeak(
    LinkageObject = SmallLinkageObject,
    gene_list = gene_list,
    genelist_idtype = "external_gene_name"
  )
peakAnno <- PeakAnnottation(LinkageObject, Species = "Homo")
}
