% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PFMatrixList}
\alias{PFMatrixList}
\title{Homo sapiens PWM matrix list of motif}
\format{
## `PFMatrixList`
}
\usage{
PFMatrixList
}
\description{
include 692 motifs
}
\details{
Report ...
}
\keyword{datasets}
