% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pwm_library_dt}
\alias{pwm_library_dt}
\title{Homo sapiens motif id and name}
\format{
## `pwm_library_dt`
A data frame with 692 rows and 2 columns:
\describe{
  \item{ID}{motif id}
  \item{name}{motif name}
}
}
\usage{
pwm_library_dt
}
\description{
Report ...
}
\keyword{datasets}
