% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegulatoryPeaks.R
\name{regulatory_peak}
\alias{regulatory_peak}
\title{Finding the regulatory peak of genes}
\usage{
regulatory_peak(RNA.seq, ATAC.seq, range, cor_method)
}
\arguments{
\item{RNA.seq}{Expression of a single or multiple genes}

\item{ATAC.seq}{ATAC-Seq Expression Matrix}

\item{range}{Expand the range of genes to find regulatory peaks}

\item{cor_method}{For correlation calculation, pearson/spearman/kendall can be selected}
}
\value{
Regulate the peak list
}
\description{
Finding the regulatory peak of genes
}
\examples{
RNA.seq.dir <- system.file("extdata","TCGA-BRCA-RNA.txt", package = "Linkage")
RNA.seq <- data.table::fread(RNA.seq.dir, header = T)[c(1:10), ]
position.dir <- system.file("extdata","homo.gene_positions.plus.txt", package = "Linkage")
position <- data.table::fread(position.dir, header = T, sep = "\t")
RNA.seq <- merge(position, RNA.seq, by.x = "ensembl_gene_id", by.y = "ensembl_gene_id")
Homo.list.files.dir <- system.file("extdata","Homo.ATAC", package = "Linkage")
Homo.list.files <- list.files(Homo.list.files.dir)
Homo.list.files <- paste0(Homo.list.files.dir,"/", Homo.list.files)
Homo.df_list <- lapply(Homo.list.files, function(file) data.table::fread(file, header = T))
ATAC.seq <- do.call(rbind, Homo.df_list)
regulatory_peak(RNA.seq, ATAC.seq, 500000, "spearman")
}
