% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegulatoryPeaks.R
\name{RegulatoryPeak}
\alias{RegulatoryPeak}
\title{Finding the regulatory peak of genes.}
\usage{
RegulatoryPeak(
  LinkageObject,
  gene_list,
  genelist_idtype,
  range = 5e+05,
  cor_method = "spearman",
  filter_col = "FDR",
  filter_value = 0.05
)
}
\arguments{
\item{LinkageObject}{An Linkage Object.}

\item{gene_list}{Character value,The gene you want to query.}

\item{genelist_idtype}{There are ensembl_gene_id, external_gene_name, entrezgene_id types of genetic IDs to choose from.}

\item{range}{Expand the range of genes to find regulatory peaks.}

\item{cor_method}{For correlation calculation, pearson/spearman/kendall can be selected.}

\item{filter_col}{p_value,FDR,rho can be selected.}

\item{filter_value}{Cutoff value of filter_col.}
}
\value{
Regulate the peak list
}
\description{
Finding the regulatory peak of genes.
}
\examples{
library(LinkageR)
data("SmallLinkageObject")
gene_list <- c("TSPAN6", "CD99", "KLHL13")
LinkageObject <-
  RegulatoryPeak(
    LinkageObject = SmallLinkageObject,
    gene_list = gene_list,
    genelist_idtype = "external_gene_name"
  )
}
