% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SASPGeneSet}
\alias{SASPGeneSet}
\title{Senescence-associated secretory phenotype Gene Set}
\format{
## `SASPGeneSet`
A character with 83 genes:
}
\usage{
SASPGeneSet
}
\description{
Report ...
}
\keyword{datasets}
