% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAnalysis.R
\name{SeqLogoPlot}
\alias{SeqLogoPlot}
\title{Transcription factor Seqlogo plot.}
\usage{
SeqLogoPlot(motif_ID)
}
\arguments{
\item{motif_ID}{Motif ID.}
}
\value{
seqLogo plot.
}
\description{
Transcription factor Seqlogo plot.
}
\examples{
library(Linkage)
SeqLogoPlot("MA0618.1")
}
