% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateLinkageObject.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,LinkageObject,ANY-method}
\title{show method}
\usage{
show(object)
}
\arguments{
\item{object}{A \code{LinkageObject} instance}
}
\value{
message
}
\description{
show method for \code{LinkageObject} instance
}
