% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Small_ATAC}
\alias{Small_ATAC}
\title{Small ATAC-seq count}
\format{
## `Small_ATAC`
A data frame with 32 rows and 75 columns:
}
\usage{
Small_ATAC
}
\description{
Small ATAC-seq count
}
\keyword{datasets}
