% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SmallLinkageObject}
\alias{SmallLinkageObject}
\title{An small LinkageObject with 5 genes and 3982 peaks}
\format{
## `SmallLinkageObject`
}
\usage{
SmallLinkageObject
}
\description{
Report ...
}
\keyword{datasets}
