% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakAnnotation.R
\name{upsetplot.csAnno}
\alias{upsetplot.csAnno}
\title{upsetplot.csAnno}
\usage{
upsetplot.csAnno(
  LinkageObject,
  order_by = "freq",
  vp = list(x = 0.6, y = 0.7, width = 0.8, height = 0.8)
)
}
\arguments{
\item{LinkageObject}{An LinkageObject after the annottation.}

\item{order_by}{How the intersections in the matrix should be ordered by. Options include frequency (entered as "freq"), degree, or both in any order.}

\item{vp}{}
}
\value{
An upsetplot after the annottation.
}
\description{
upsetplot.csAnno
}
\examples{
data("LinkageObject")
gene_list <- c("TSPAN6", "CD99", "KLHL13")
LinkageObject <- regulatory_peak(LinkageObject = LinkageObject, gene_list = gene_list, genelist_idtype = "external_gene_name")
peakAnno <- Peak_Annottation(LinkageObject, Species = "Homo")
upsetplot.csAnno(LinkageObject = peakAnno)
}
