% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.KEGG.Enrichment.R
\name{upsetplot}
\alias{upsetplot}
\title{Enrichment analysis Upsetplot}
\usage{
upsetplot(object, ...)
}
\arguments{
\item{object}{Go or KEGG object}

\item{...}{}
}
\value{
Enrichment upset plot.
}
\description{
Enrichment analysis Upsetplot
}
\examples{
extdatadir <- system.file(paste0("extdata"), package = "Linkage")
gene_list <- read.table(paste0(extdatadir,"/Senescence-associated secretory phenotype.txt"))
go <- GO.enrichment(gene_list = gene_list$V1, Species = "Homo", genelist_idtype = "external_gene_name")
upsetplot(go)
}
