% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakAnnotation.R
\name{vennpie}
\alias{vennpie}
\title{Annotation regulatory peak upset plot}
\usage{
vennpie(object, vennpie = TRUE, ...)
}
\arguments{
\item{object}{A Annottation object.}

\item{vennpie}{logical value, whether to include vennpie.The default is TRUE}

\item{...}{...}
}
\value{
A upset plot
}
\description{
Annotation regulatory peak upset plot
}
\examples{
dir <- system.file("extdata","ENSG00000000419.csv", package = "Linkage")
peakfile <- read.csv(dir, header = TRUE)
peakAnno <- Peak_Annottation(peakfile, Species = "Homo")
vennpie(peakAnno)
}
