% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNetwork.R
\name{VisualGRNs}
\alias{VisualGRNs}
\title{VisualGRNs.}
\usage{
VisualGRNs(LinkageObject)
}
\arguments{
\item{LinkageObject}{An Linkage Object after FilterGRNs .}
}
\value{
An visNetwork
}
\description{
VisualGRNs.
}
\examples{
library(LinkageR)
data("Small_ATAC.rda")
data("Small_RNA.rda")
LinkageObject <-
  CreateLinkageObject(
    ATAC_count = Small_ATAC,
    RNA_count = Small_RNA,
    Species = "Homo",
    id_type = "ensembl_gene_id"
  )
gene_list <- c("TSPAN6", "CD99", "KLHL13")
LinkageObject <-
  RegulatoryPeak(
    LinkageObject = LinkageObject,
    gene_list = gene_list,
    genelist_idtype = "external_gene_name"
  )
LinkageObject <-
  BuildGRNs(LinkageObject = LinkageObject,
                        Species = "Homo",
                        TF_cor_method = "pearson")
LinkageObject <-
  FilterGRNs(
    LinkageObject = LinkageObject,
    genelist_idtype = "entrezgene_id",
    filter_col = "FDR",
    filter_value = 0.01
  )
VisualGRNs(LinkageObject)
}
