% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.KEGG.Enrichment.R
\name{wordcloud}
\alias{wordcloud}
\title{Enrichment analysis Wordcloud}
\usage{
wordcloud(object, Type, plotly = TRUE, ...)
}
\arguments{
\item{object}{Go or KEGG object.}

\item{Type}{Indicate whether it is a GO or KEGG object, GO and KEGG can be selected.}

\item{plotly}{logical value,Whether or not to use plotly.The default is TRUE.}

\item{...}{...}
}
\value{
Enrichment wordcloud.
}
\description{
Enrichment analysis Wordcloud
}
\examples{
extdatadir <- system.file(paste0("extdata"), package = "Linkage")
gene_list <- read.table(paste0(extdatadir, "/Senescence-associated secretory phenotype.txt"))
go <- GO.enrichment(gene_list = gene_list$V1, Species = "Homo", genelist_idtype = "external_gene_name")
wordcloud(go, Type = "GO")
}
