% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.KEGG.Enrichment.R
\name{barplot}
\alias{barplot}
\title{Enrichment analysis Barplot}
\usage{
barplot(object, plotly = TRUE, ...)
}
\arguments{
\item{object}{Go or KEGG object.}

\item{plotly}{logical value,Whether or not to use plotly.The default is TRUE.}

\item{...}{...}
}
\value{
Enrichment bar plot.
}
\description{
Enrichment analysis Barplot
}
\examples{
extdatadir <- system.file("extdata", "Senescence-associated secretory phenotype.txt", package = "Linkage")
gene_list <- read.table(extdatadir)
go <- GO.enrichment(gene_list = gene_list$V1, Species = "Homo", genelist_idtype = "external_gene_name")
barplot(go)
}
