% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakVisualization.R
\name{box_plot}
\alias{box_plot}
\title{Regulate peak genomic location browsing boxplot}
\usage{
box_plot(peakfile, gene, select_peak, plotly)
}
\arguments{
\item{peakfile}{ATAC-Seq expression matrix or bed file.}

\item{gene}{Genes regulated by peak.}

\item{select_peak}{The line index of ATAC-Seq expression matrix or bed file.}

\item{plotly}{logical value,Whether or not to use plotly.The default is TRUE.}

\item{...}{...}
}
\value{
A box plot.
}
\description{
Regulate peak genomic location browsing boxplot
}
\examples{
peakfile.dir <- system.file("extdata","ENSG00000000419.csv", package = "Linkage")
peakfile <- read.csv(peakfile.dir, header = T,check.names = F)
gene.dir <- system.file("extdata","TCGA-BRCA-RNA.txt", package = "Linkage")
gene <- data.table::fread(gene.dir, header = T,check.names = F)
gene <- gene[gene$ensembl_gene_id == "ENSG00000000419", ]
p <- trackplot(peakfile = peakfile, 1, "Homo")
box_plot(peakfile, gene, select_peak = 1, F)
}
