% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNetwork.R
\name{BuildNetwork}
\alias{BuildNetwork}
\title{BuildNetwork.}
\usage{
BuildNetwork(LinkageObject)
}
\arguments{
\item{LinkageObject}{An Linkage Object after CreateNetworkObject.}
}
\value{
An visNetwork
}
\description{
BuildNetwork.
}
\examples{
library(linkage)
library(LinkageData)
ATAC.seq <- BreastCancerATAC()
RNA.seq <- BreastCancerRNA()
LinkageObject <-
  CreateLinkageObject(
    ATAC_count = ATAC.seq,
    RNA_count = RNA.seq,
    Species = "Homo",
    id_type = "ensembl_gene_id"
  )
gene_list <- c("TSPAN6", "CD99", "KLHL13")
LinkageObject <-
  RegulatoryPeak(
    LinkageObject = LinkageObject,
    gene_list = gene_list,
    genelist_idtype = "external_gene_name"
  )
LinkageObject <-
  MultipleMotifAnalysis(LinkageObject = LinkageObject,
                        Species = "Homo",
                        TF_cor_method = "pearson")
LinkageObject <-
  CreateNetworkObject(
    LinkageObject = LinkageObject,
    genelist_idtype = "entrezgene_id",
    filter_col = "FDR",
    filter_value = 0.01
  )
BuildNetwork(LinkageObject)
}
