% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.KEGG.Enrichment.R
\name{cnetplot}
\alias{cnetplot}
\title{Enrichment analysis Cnetplot}
\usage{
cnetplot(object, ...)
}
\arguments{
\item{object}{Go or KEGG object}

\item{...}{...}
}
\value{
Enrichment network.
}
\description{
Enrichment analysis Cnetplot
}
\examples{
gene_list <- read.table("inst/extdata/Senescence-associated secretory phenotype.txt")
go <- GO.enrichment(gene_list = gene_list$V1, Species = "Homo", genelist_idtype = "external_gene_name")
cnetplot(go)
}
