% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNetwork.R
\name{CreateNetworkObject}
\alias{CreateNetworkObject}
\title{Create Network Object.}
\usage{
CreateNetworkObject(LinkageObject, genelist_idtype, filter_col, filter_value)
}
\arguments{
\item{LinkageObject}{An Linkage Object after MultipleMotifAnalysis.}

\item{genelist_idtype}{There are ensembl_gene_id, external_gene_name, entrezgene_id types of genetic IDs to choose from.}

\item{filter_col}{p_value,FDR,rho can be selected.}

\item{filter_value}{Cutoff value of filter_col.}
}
\value{
Network Object.
}
\description{
Create Network Object.
}
\examples{
library(Linkage)
library(LinkageData)
ATAC.seq <- BreastCancerATAC()
RNA.seq <- BreastCancerRNA()
LinkageObject <-
  CreateLinkageObject(
    ATAC_count = ATAC.seq,
    RNA_count = RNA.seq,
    Species = "Homo",
    id_type = "ensembl_gene_id"
  )
gene_list <- c("TSPAN6", "CD99", "KLHL13")
LinkageObject <-
  RegulatoryPeak(
    LinkageObject = LinkageObject,
    gene_list = gene_list,
    genelist_idtype = "external_gene_name"
  )
LinkageObject <-
  MultipleMotifAnalysis(LinkageObject = LinkageObject,
                        Species = "Homo",
                        TF_cor_method = "pearson")
LinkageObject <-
  CreateNetworkObject(
    LinkageObject = LinkageObject,
    genelist_idtype = "entrezgene_id",
    filter_col = "FDR",
    filter_value = 0.01
  )
}
