% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateLinkageObject.R
\docType{class}
\name{LinkageObject-class}
\alias{LinkageObject-class}
\title{The LinkageObject class}
\value{
Linkage Object
}
\description{
The LinkageObject class
}
\section{Slots}{

\describe{
\item{\code{RNA.mtrix}}{data.frame.}

\item{\code{ATAC.matrix}}{data.frame.}

\item{\code{active.gene}}{data.frame.}

\item{\code{cor.peak}}{list.}

\item{\code{cor.peak.annotation}}{list.}

\item{\code{Motif}}{list.}

\item{\code{Gene_TF}}{data.frame.}

\item{\code{geneid}}{character.}

\item{\code{NetworkSetting}}{list.}
}}

