% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAnalysis.R
\name{motif_analysis}
\alias{motif_analysis}
\title{Regulatory peak motif enrichment analysis}
\usage{
motif_analysis(peakfile, Species)
}
\arguments{
\item{peakfile}{ATAC-Seq expression matrix or bed file}

\item{Species}{Select the species, Homo or Mus.}
}
\value{
Motif result
}
\description{
Regulatory peak motif enrichment analysis
}
\examples{
extdatadir <- system.file(paste0("extdata"), package = "Linkage")
peakfile <- read.csv(paste0(extdatadir,"/ENSG00000000419.csv"), header = T)
motif_analysis(peakfile, "Homo")
}
