% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAnalysis.R
\name{MultipleMotifAnalysis}
\alias{MultipleMotifAnalysis}
\title{MotifAnalysis.}
\usage{
MultipleMotifAnalysis(LinkageObject, Species = "Homo", TF_cor_method)
}
\arguments{
\item{LinkageObject}{An Linkage Object after RegulatoryPeak.}

\item{Species}{Select the species, Homo or Mus.The default is Homo.}

\item{TF_cor_method}{For correlation calculation, pearson/spearman/kendall can be selected.}
}
\value{
An Linkage Object after MultipleMotifAnalysis
}
\description{
MotifAnalysis.
}
\examples{
library(Linkage)
library(LinkageData)
ATAC.seq <- BreastCancerATAC()
RNA.seq <- BreastCancerRNA()
LinkageObject <-
  CreateLinkageObject(
    ATAC_count = ATAC.seq,
    RNA_count = RNA.seq,
    Species = "Homo",
    id_type = "ensembl_gene_id"
  )
gene_list <- c("TSPAN6", "CD99", "KLHL13")
LinkageObject <-
  RegulatoryPeak(
    LinkageObject = LinkageObject,
    gene_list = gene_list,
    genelist_idtype = "external_gene_name"
  )
LinkageObject <-
  MultipleMotifAnalysis(LinkageObject = LinkageObject,
                        Species = "Homo",
                        TF_cor_method = "pearson")
}
