% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakAnnotation.R
\name{Peak_Annottation}
\alias{Peak_Annottation}
\title{Annotation regulatory peak}
\usage{
Peak_Annottation(peakfile, Species = "Homo", ...)
}
\arguments{
\item{peakfile}{ATAC-Seq expression matrix or bed file.}

\item{Species}{Select the species, Homo or Mus.The default is Homo.}

\item{...}{}
}
\value{
A Annottation object
}
\description{
Annotation regulatory peak
}
\examples{
extdatadir <- system.file(paste0("extdata"), package = "Linkage")
peakfile <- read.csv(paste0(extdatadir,"/ENSG00000000419.csv"), header = T)
peakAnno <- Peak_Annottation(peakfile, Species = "Homo")
}
