% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakAnnotation.R
\name{Peak_Annottation}
\alias{Peak_Annottation}
\title{Annotation regulatory peak}
\usage{
Peak_Annottation(LinkageObject, Species = "Homo", ...)
}
\arguments{
\item{LinkageObject}{An Linkage Object after regulatory_peak.}

\item{Species}{Select the species, Homo or Mus.The default is Homo.}

\item{...}{...}
}
\value{
An LinkageObject after the annottation.
}
\description{
Annotation regulatory peak
}
\examples{
data("LinkageObject")
gene_list <- c("TSPAN6", "CD99", "KLHL13")
LinkageObject <- regulatory_peak(LinkageObject = LinkageObject, gene_list = gene_list, genelist_idtype = "external_gene_name")
peakAnno <- Peak_Annottation(LinkageObject, Species = "Homo")
}
