% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegulatoryPeaks.R
\name{regulatory_peak}
\alias{regulatory_peak}
\title{Finding the regulatory peak of genes}
\usage{
regulatory_peak(RNA.seq, ATAC.seq, range, cor_method)
}
\arguments{
\item{RNA.seq}{Expression of a single or multiple genes}

\item{ATAC.seq}{ATAC-Seq Expression Matrix}

\item{range}{Expand the range of genes to find regulatory peaks}

\item{cor_method}{For correlation calculation, pearson/spearman/kendall can be selected}
}
\value{
Regulate the peak list
}
\description{
Finding the regulatory peak of genes
}
\examples{
extdatadir <- system.file(paste0("extdata"), package = "Linkage")
RNA.seq <- data.table::fread(paste0(extdatadir,"/TCGA-BRCA-RNA.txt"), header = T)[c(1:10), ]
position <- data.table::fread(paste0(extdatadir,"/homo.gene_positions.plus.txt"), header = T, sep = "\t")
RNA.seq <- merge(position, RNA.seq, by.x = "ensembl_gene_id", by.y = "ensembl_gene_id")
Homo.list.files <- list.files(paste0(extdatadir,"/Homo.ATAC/"))
Homo.list.files <- paste0(paste0(extdatadir,"/Homo.ATAC/"),Homo.list.files)
Homo.df_list <- lapply(Homo.list.files,function(file) data.table::fread(file,header = T))
ATAC.seq <- do.call(rbind, Homo.df_list)
regulatory_peak(RNA.seq, ATAC.seq, 500000, "spearman")
}
