% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAnalysis.R
\name{seqLogo_plot}
\alias{seqLogo_plot}
\title{Transcription factor Seqlogo plot.}
\usage{
seqLogo_plot(motif_ID)
}
\arguments{
\item{motif_ID}{Motif ID.}
}
\value{
seqLogo plot.
}
\description{
Transcription factor Seqlogo plot.
}
\examples{
seqLogo_plot("MA0618.1")
}
