% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakVisualization.R
\name{Trackplot}
\alias{Trackplot}
\title{Plot Trackplot.Users initially select a regulatory peak that obtained from the Regulatory Peaks Detection Module. Linkage then categorizes samples into five groups based on the quantitative chromatin accessibility of the specific regulatory peak, ranging from low to high for each individual sample. The coverage track of mapped ATAC-seq reads and the boxplot of the target gene for each group will be shown simultaneously.}
\usage{
Trackplot(LinkageObject, Geneid, peakid, Species = "Homo")
}
\arguments{
\item{LinkageObject}{An Linkage Object after regulatory_peak.}

\item{Geneid}{The gene you want to query.}

\item{peakid}{The peak you want to query.}

\item{Species}{Select the species, Homo or Mus.The default is Homo.}
}
\value{
Trackplot.
}
\description{
Plot Trackplot.Users initially select a regulatory peak that obtained from the Regulatory Peaks Detection Module. Linkage then categorizes samples into five groups based on the quantitative chromatin accessibility of the specific regulatory peak, ranging from low to high for each individual sample. The coverage track of mapped ATAC-seq reads and the boxplot of the target gene for each group will be shown simultaneously.
}
\examples{
data("LinkageObject")
gene_list <- c("TSPAN6", "CD99", "KLHL13")
LinkageObject <- regulatory_peak(LinkageObject = LinkageObject, gene_list = gene_list, genelist_idtype = "external_gene_name")
Trackplot(LinkageObject,Geneid = "TSPAN6",peakid = "chrX:100635908-100636408",Species = "Homo")
}
