% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakVisualization.R
\name{TrackPlot}
\alias{TrackPlot}
\title{TrackPlot}
\usage{
TrackPlot(LinkageObject, Geneid, peakid, Species = "Homo")
}
\arguments{
\item{LinkageObject}{An Linkage Object after RegulatoryPeak.}

\item{Geneid}{The gene you want to query.}

\item{peakid}{The peak you want to query.}

\item{Species}{Select the species, Homo or Mus.The default is Homo.}
}
\value{
TrackPlot.
}
\description{
TrackPlot
}
\examples{
library(Linkage)
data("SmallLinkageObject")
gene_list <- c("TSPAN6", "CD99", "KLHL13")
  LinkageObject <-
RegulatoryPeak(
  LinkageObject = SmallLinkageObject,
  gene_list = gene_list,
  genelist_idtype = "external_gene_name"
)
TrackPlot(
  LinkageObject,
  Geneid = "TSPAN6",
  peakid = "chrX:100635908-100636408",
  Species = "Homo"
)
}
