% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakVisualization.R
\name{trackplot}
\alias{trackplot}
\title{Regulate peak genomic location browsing}
\usage{
trackplot(peakfile, select_peak, Species)
}
\arguments{
\item{peakfile}{ATAC-Seq expression matrix or bed file.}

\item{select_peak}{The line index of ATAC-Seq expression matrix or bed file.}

\item{Species}{Select the species, Homo or Mus.The default is Homo.}
}
\value{
A track plot
}
\description{
Regulate peak genomic location browsing
}
\examples{
peakfile.dir <- system.file("extdata","ENSG00000000419.csv", package = "Linkage")
peakfile <- read.csv(peakfile.dir, header = T,check.names = F)
gene.dir <- system.file("extdata","TCGA-BRCA-RNA.txt", package = "Linkage")
gene <- data.table::fread(gene.dir, header = T,check.names = F)
gene <- gene[gene$ensembl_gene_id == "ENSG00000000419", ]
p <- trackplot(peakfile = peakfile, 1, "Homo")
box_plot(peakfile, gene, select_peak = 1, F)
}
