% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNetwork.R
\name{VisualGRNs}
\alias{VisualGRNs}
\title{VisualGRNs.}
\usage{
VisualGRNs(LinkageObject)
}
\arguments{
\item{LinkageObject}{An Linkage Object after FilterGRNs .}
}
\value{
An visNetwork
}
\description{
VisualGRNs.
}
\examples{
library(linkage)
library(LinkageData)
ATAC.seq <- BreastCancerATAC()
RNA.seq <- BreastCancerRNA()
LinkageObject <-
  CreateLinkageObject(
    ATAC_count = ATAC.seq,
    RNA_count = RNA.seq,
    Species = "Homo",
    id_type = "ensembl_gene_id"
  )
gene_list <- c("TSPAN6", "CD99", "KLHL13")
LinkageObject <-
  RegulatoryPeak(
    LinkageObject = LinkageObject,
    gene_list = gene_list,
    genelist_idtype = "external_gene_name"
  )
LinkageObject <-
  BuildGRNs(LinkageObject = LinkageObject,
                        Species = "Homo",
                        TF_cor_method = "pearson")
LinkageObject <-
  BuildGRNs(
    LinkageObject = LinkageObject,
    genelist_idtype = "entrezgene_id",
    filter_col = "FDR",
    filter_value = 0.01
  )
VisualGRNs(LinkageObject)
}
